backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2

# LLM & Agents
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5
langgraph==0.2.45
openai==1.54.3

# Vector DB & Embeddings
chromadb==0.5.15
sentence-transformers==3.2.1

# Medical NLP
scispacy==0.5.5
spacy==3.7.6

# Data Processing
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2

# Document Parsing
pypdf==5.0.1
fpdf2==2.8.1
python-docx==1.1.2

# Visualization & Reports
matplotlib==3.9.2
plotly==5.24.1

# CORS & HTTP
httpx==0.27.2
python-multipart==0.0.12

# Date handling
python-dateutil==2.9.0
